requirements.txt
asyncpg
sqlalchemy
alembic
python-dotenv
pydantic
greenlet
sqlalchemy[asyncio]
psycopg2-binary
pytest
fastapi
httpx
uvicorn
pytest-asyncio
python-jose==3.3.0
ujson==5.10.0
passlib==1.7.4
bcrypt==3.2.0
pydantic[email]
python-multipart
fastapi-pagination
sqladmin==0.18.0
itsdangerous
aioredis
fastapi-cache2
redis